RMB version update, package naming update
