+++ pom.xml
@@ -14,7 +14,7 @@
   </licenses>
 
   <properties>
-    <raml-module-builder.version>29.3.0</raml-module-builder.version>
+    <raml-module-builder.version>29.3.1</raml-module-builder.version>
     <vertx.version>3.8.4</vertx.version>
     <junit.version>4.12</junit.version>
     <rest-assured.version>3.3.0</rest-assured.version>
