+++ src/main/java/org/folio/rest/impl/DataExportImplFileDefinitionImpl.java
@@ -21,8 +21,8 @@ import org.folio.rest.exceptions.HttpException;
 import org.folio.rest.jaxrs.model.FileDefinition;
 import org.folio.rest.jaxrs.resource.DataExportFileDefinitions;
 import org.folio.rest.tools.utils.TenantTool;
-import org.folio.service.upload.FileUploadService;
-import org.folio.service.upload.definition.FileDefinitionService;
+import org.folio.service.file.upload.FileUploadService;
+import org.folio.service.file.definition.FileDefinitionService;
 import org.folio.spring.SpringContextUtil;
 import org.folio.util.ErrorCodes;
 import org.folio.util.ExceptionToResponseMapper;
