+++ src/main/java/org/folio/rest/impl/InitAPIImpl.java
@@ -11,8 +11,8 @@ import java.lang.invoke.MethodHandles;
 import java.util.Properties;
 import org.folio.config.ApplicationConfig;
 import org.folio.rest.resource.interfaces.InitAPI;
-import org.folio.service.manager.exportmanager.ExportManager;
-import org.folio.service.manager.inputdatamanager.InputDataManager;
+import org.folio.service.manager.export.ExportManager;
+import org.folio.service.manager.input.InputDataManager;
 import org.folio.spring.SpringContextUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
