+++ src/main/java/org/folio/service/export/LocalFileSystemExportService.java
@@ -8,7 +8,7 @@ import java.util.List;
 import org.apache.commons.collections4.CollectionUtils;
 import org.folio.rest.jaxrs.model.FileDefinition;
 import org.folio.service.export.storage.ExportStorageService;
-import org.folio.service.upload.storage.FileStorage;
+import org.folio.service.file.storage.FileStorage;
 import org.marc4j.MarcJsonReader;
 import org.marc4j.MarcReader;
 import org.marc4j.MarcStreamWriter;
