+++ src/main/java/org/folio/service/manager/export/ExportResult.java
@@ -1,0 +1,5 @@
+package org.folio.service.manager.export;
+
+public enum ExportResult {
+  COMPLETED, IN_PROGRESS, ERROR;
+}
