+++ src/main/java/org/folio/service/manager/export/package-info.java
@@ -1,0 +1,4 @@
+@ModuleGen(groupPackage = "org.folio.service.manager.export", name = "export-manager")
+package org.folio.service.manager.export;
+
+import io.vertx.codegen.annotations.ModuleGen;
