--- src/main/java/org/folio/service/manager/exportmanager/package-info.java
@@ -1,4 +1,0 @@
-@ModuleGen(groupPackage = "org.folio.service.manager.exportmanager", name = "export-manager")
-package org.folio.service.manager.exportmanager;
-
-import io.vertx.codegen.annotations.ModuleGen;
