--- src/main/java/org/folio/service/manager/exportresult/ExportResult.java
@@ -1,5 +1,0 @@
-package org.folio.service.manager.exportresult;
-
-public enum ExportResult {
-  COMPLETED, IN_PROGRESS, ERROR;
-}
