+++ src/main/java/org/folio/service/manager/input/InputDataContext.java
@@ -1,0 +1,21 @@
+package org.folio.service.manager.input;
+
+import io.vertx.core.shareddata.Shareable;
+import org.folio.service.file.reader.SourceReader;
+
+public class InputDataContext implements Shareable {
+
+  private SourceReader sourceReader;
+
+  public InputDataContext(SourceReader sourceReader) {
+    this.sourceReader = sourceReader;
+  }
+
+  public SourceReader getSourceReader() {
+    return sourceReader;
+  }
+
+  public void setSourceReader(SourceReader sourceReader) {
+    this.sourceReader = sourceReader;
+  }
+}
