+++ src/main/java/org/folio/service/manager/input/package-info.java
@@ -1,0 +1,4 @@
+@ModuleGen(groupPackage = "org.folio.service.manager.input", name = "input-data-manager")
+package org.folio.service.manager.input;
+
+import io.vertx.codegen.annotations.ModuleGen;
