--- src/main/java/org/folio/service/manager/inputdatamanager/package-info.java
@@ -1,4 +1,0 @@
-@ModuleGen(groupPackage = "org.folio.service.manager.inputdatamanager", name = "input-data-manager")
-package org.folio.service.manager.inputdatamanager;
-
-import io.vertx.codegen.annotations.ModuleGen;
