--- src/main/java/org/folio/service/upload/storage/FileStorage.java
@@ -1,25 +1,0 @@
-package org.folio.service.upload.storage;
-
-import io.vertx.core.Future;
-import org.folio.rest.jaxrs.model.FileDefinition;
-
-/**
- * File storage service, service to save files.
- */
-public interface FileStorage {
-
-  /**
-   * Saves bytes to the storage asynchronously
-   */
-  Future<FileDefinition> saveFileDataAsync(byte[] data, FileDefinition fileDefinition);
-
-  /**
-   * Save bytes to the storage in blocking manner
-   */
-  FileDefinition saveFileDataBlocking(byte[] data, FileDefinition fileDefinition);
-
-  /**
-   * Deletes File and related parent directory from the storage and returns true if succeeded
-   */
-  Future<Boolean> deleteFileAndParentDirectory(FileDefinition fileDefinition);
-}
